+++ src/components/Home/homeNews.ts
@@ -2,7 +2,7 @@ import { defineComponent, computed } from "@vue/composition-api"
 import Icon from "vue-awesome/components/Icon.vue"
 import Loader from "@/components/Loader.vue"
 import AppUrls from "@/utils/urls"
-import useNews from "@/factories/useNewsFactory"
+import useNews from "@/factories/useNews"
 import { AsyncDataStatus } from "@/factories/useAsyncData"
 
 const HomeNews = defineComponent({
